requirements.txt
Flask==1.0.2
pandas==0.24.0
requests==2.21.0
mysql-connector==2.1.6